Brought back the page header and make it display tab buttons to navigate easy to the GUI pages. Also fixed a bug in the column breaking code.
Warning: The pretty tabs required a little hack. In the column breaking code within a gui page the x-offset of the new column is defined by the width of the last element in the old column, not by the grid size anymore. So please make sure all the element within a page have the same width (this is default).